+++ src/Jenssegers/Mongodb/Eloquent/Builder.php
@@ -35,14 +35,6 @@ class Builder extends EloquentBuilder
      */
     public function update(array $values, array $options = [])
     {
-        // Intercept operations on embedded models and delegate logic
-        // to the parent relation instance.
-        if ($relation = $this->model->getParentRelation()) {
-            $relation->performUpdate($this->model, $values);
-
-            return 1;
-        }
-
         return $this->toBase()->update($this->addUpdatedAtColumn($values), $options);
     }
 
@@ -51,14 +43,6 @@ class Builder extends EloquentBuilder
      */
     public function insert(array $values)
     {
-        // Intercept operations on embedded models and delegate logic
-        // to the parent relation instance.
-        if ($relation = $this->model->getParentRelation()) {
-            $relation->performInsert($this->model, $values);
-
-            return true;
-        }
-
         return parent::insert($values);
     }
 
@@ -67,14 +51,6 @@ class Builder extends EloquentBuilder
      */
     public function insertGetId(array $values, $sequence = null)
     {
-        // Intercept operations on embedded models and delegate logic
-        // to the parent relation instance.
-        if ($relation = $this->model->getParentRelation()) {
-            $relation->performInsert($this->model, $values);
-
-            return $this->model->getKey();
-        }
-
         return parent::insertGetId($values, $sequence);
     }
 
@@ -83,14 +59,6 @@ class Builder extends EloquentBuilder
      */
     public function delete()
     {
-        // Intercept operations on embedded models and delegate logic
-        // to the parent relation instance.
-        if ($relation = $this->model->getParentRelation()) {
-            $relation->performDelete($this->model);
-
-            return $this->model->getKey();
-        }
-
         return parent::delete();
     }
 
@@ -99,23 +67,6 @@ class Builder extends EloquentBuilder
      */
     public function increment($column, $amount = 1, array $extra = [])
     {
-        // Intercept operations on embedded models and delegate logic
-        // to the parent relation instance.
-        if ($relation = $this->model->getParentRelation()) {
-            $value = $this->model->{$column};
-
-            // When doing increment and decrements, Eloquent will automatically
-            // sync the original attributes. We need to change the attribute
-            // temporary in order to trigger an update query.
-            $this->model->{$column} = null;
-
-            $this->model->syncOriginalAttribute($column);
-
-            $result = $this->model->update([$column => $value]);
-
-            return $result;
-        }
-
         return parent::increment($column, $amount, $extra);
     }
 
@@ -124,21 +75,6 @@ class Builder extends EloquentBuilder
      */
     public function decrement($column, $amount = 1, array $extra = [])
     {
-        // Intercept operations on embedded models and delegate logic
-        // to the parent relation instance.
-        if ($relation = $this->model->getParentRelation()) {
-            $value = $this->model->{$column};
-
-            // When doing increment and decrements, Eloquent will automatically
-            // sync the original attributes. We need to change the attribute
-            // temporary in order to trigger an update query.
-            $this->model->{$column} = null;
-
-            $this->model->syncOriginalAttribute($column);
-
-            return $this->model->update([$column => $value]);
-        }
-
         return parent::decrement($column, $amount, $extra);
     }