+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Removed
+- EmbedsOne and EmbedsMany relations by [@divine](https://github.com/divine).